feat: Add 18% GST option for store settings
This commit adds an 18% GST option to the store settings for both takeaway and dine-in orders. This allows administrators to configure the GST rate to 18% in addition to the existing 5% and 12% options. The change is implemented by adding a new `<ng-option>` with the value "18" to the GST selection dropdowns in the store settings component.